+++ example/lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:example/src/configuration/default_application_configuration.dart';
-import 'package:flutter/material.dart';
 import 'package:fappconfiguration/fappconfiguration.dart';
+import 'package:flutter/material.dart';
 
 void main() {
   runApp(