+++ example/lib/src/ids.dart
@@ -1,2 +1,5 @@
 /// Unique feature identifier for an incrementer.
 const String incrementerFeature = "incrementer";
+
+/// Unique feature identifier for theme demo.
+const String dependencyExample = "dependency_example";